refactor: Move installLauncherPR to store submodule
